+++ src/main/java/webserver/handler/urlhandler/LogoutHandler.java
@@ -1,4 +1,4 @@
-package webserver.handler.url;
+package webserver.handler.urlhandler;
 
 import config.Config;
 import db.SessionStore;
@@ -8,10 +8,9 @@ import webserver.StatusCode;
 import webserver.Url;
 import webserver.constants.Constants;
 import webserver.handler.UrlHandler;
-import webserver.handler.error.ErrorHandler;
+import webserver.handler.errorhandler.ErrorHandler;
 import webserver.request.HttpRequest;
 import webserver.response.HttpResponse;
-import webserver.utils.HttpRequestUtils;
 import webserver.utils.HttpResponseUtils;
 import webserver.utils.SidUtils;
